apps/cli/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kube-borg-backup-cli"
version = "6.0.0"
description = "CLI tool for kube-borg-backup restore operations"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "kubernetes>=28.0.0",
    "PyYAML>=6.0",
]

[project.scripts]
kbb = "kbb.main:main"
